add: After changed target battery, change battery

+++ app/src/main/java/com/eddy/nrf/presentation/ui/BikeUiState.kt
@@ -5,6 +5,6 @@ data class BikeUiState(
     var distance: Float =190.0f,
     var speed: Float = 0.0f,
     var gear: Int = 1,
-    var battery: Int = 100,
+    var battery: Float = 100f,
     var proportionalFactor: Float = 1.0f
 )
